refactor(line-repository): 클래스가 너무 길어져서 해당 메소드 별로 테스트 분리
