+++ src/test/java/subway/domain/LineRepositoryTest.java
@@ -50,172 +50,4 @@ class LineRepositoryTest {
         assertThatIllegalArgumentException().isThrownBy(callable)
                 .withMessage(LineRepository.DUPLICATE_NAME_ERROR, "1호선");
     }
-
-    @Test
-    @DisplayName("존재하는 노선 삭제")
-    public void deleteStation_ExistLine_EmptyLines() {
-
-        // when
-        LineRepository newLineRepository = lineRepository.deleteLineByName("1호선");
-
-        //then
-        assertThat(newLineRepository.lines().isEmpty()).isTrue();
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 노선 삭제 시 예외 발생")
-    public void deleteStation_DoesNotExistLine_ExceptionThrown() {
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.deleteLineByName("신분당선");
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(LineRepository.DOES_NOT_EXIST_ERROR, "신분당선");
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 노선에 역 삽입 시 예외 발생")
-    public void insertStation_DoesNotExistLine_ExceptionThrown() {
-
-        // given
-        int index = 1;
-        String stationName = "잠실역";
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.insertStation("11호선", index, stationName);
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(LineRepository.DOES_NOT_EXIST_ERROR, "11호선");
-    }
-
-    @Test
-    @DisplayName("구간에 중복되지 않은 역 삽입")
-    public void insertStation_NewStation_LineRepositoryInsertedStation() {
-
-        // given
-        int index = 1;
-        String stationName = "봉천역";
-
-        // when
-        LineRepository newLineRepository =
-                lineRepository.insertStation(line.getName(), index, stationName);
-
-        //then
-        assertThat(newLineRepository.lines().size()).isEqualTo(1);
-        assertThat(newLineRepository.contains(stationName)).isTrue();
-    }
-
-    @Test
-    @DisplayName("구간에 중복되지 않은 역 삽입 시 예외 발생")
-    public void insertStation_DuplicateStation_ExceptionThrown() {
-
-        // given
-        int index = 1;
-        String stationName = "잠실역";
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.insertStation(line.getName(), index, stationName);
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(StationRepository.DUPLICATE_NAME_ERROR, stationName);
-    }
-
-    @Test
-    @DisplayName("구간 순서가 최소값보다 작을 경우 예외 발생")
-    public void insertStation_LowerThanMinimumIndex_ExceptionThrown() {
-
-        // given
-        int index = 0;
-        String stationName = "잠실역";
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.insertStation(line.getName(), index, stationName);
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(StationRepository.OUT_OF_BOUNDS_ERROR, 2);
-    }
-
-    @Test
-    @DisplayName("구간 순서가 최대값보다 클 경우 예외 발생")
-    public void insertStation_HigherThanStationSize_ExceptionThrown() {
-
-        // given
-        int index = 2;
-        String stationName = "잠실역";
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.insertStation(line.getName(), index, stationName);
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(StationRepository.OUT_OF_BOUNDS_ERROR, 2);
-    }
-
-    @Test
-    @DisplayName("노선에 등록된 역 제거")
-    public void deleteStation_oldStation_LineDeletedStation() {
-
-        // given
-        lineRepository = lineRepository.insertStation(line.getName(), 1, "봉천역");
-
-        // when
-        lineRepository = lineRepository.deleteStation("1호선", "봉천역");
-
-        //then
-        assertThat(lineRepository.lines().get(0).getStations().stations())
-                .extracting(Station::getName).containsExactly("강남역", "잠실역");
-    }
-
-    @Test
-    @DisplayName("노선의 역 개수가 최소 역 개수 이하일 경우 역 제거 시 예외 발생")
-    public void deleteStation_LessThanMinimumStationSize_ExceptionThrown() {
-
-        // when
-        ThrowableAssert.ThrowingCallable callable =
-                () -> lineRepository.deleteStation("1호선", "강남역");
-
-        //then
-        assertThatIllegalArgumentException().isThrownBy(callable)
-                .withMessage(StationRepository.TOO_LESS_STATIONS_ERROR,
-                        StationRepository.MINIMUM_STATION_SIZE);
-    }
-
-    @Test
-    @DisplayName("상행 종점을 제거할 경우 그 다음 역이 종점")
-    public void deleteStation_StartStation_StartsWithNextStation() {
-
-        // given
-        lineRepository = lineRepository.insertStation(line.getName(), 1, "봉천역");
-
-        // when
-        lineRepository = lineRepository.deleteStation("1호선", "강남역");
-
-        //then
-        assertThat(lineRepository.lines().get(0).getStations().stations())
-                .extracting(Station::getName).containsExactly("봉천역", "잠실역");
-    }
-
-    @Test
-    @DisplayName("하행 종점을 제거할 경우 그 다음 역이 종점")
-    public void deleteStation_FinalStation_EndWithNextStation() {
-
-        // given
-        lineRepository = lineRepository.insertStation(line.getName(), 1, "봉천역");
-
-        // when
-        lineRepository = lineRepository.deleteStation("1호선", "잠실역");
-
-        //then
-        assertThat(lineRepository.lines().get(0).getStations().stations())
-                .extracting(Station::getName).containsExactly("강남역", "봉천역");
-    }
 }
